Fix a memory leak in arrangement size logging (#31868)
This fixes a memory leak in arrangement size logging, which was
inadvertently introduced when combining the logging dataflows into a
single
dataflow and adding streams to share information between different
fragments. The problem is that the data needs to be processed in order,
i.e., the handler needs to see the announcement of a new arrangement
followed by updates, and at the end an indication that it was dropped.
Passing data only through the log infrastructure ensures a total order,
as
it maintains FIFO semantics. Adding additional channels to the mix
removed
the property that data arrives in order, rendering the log analysis
incorrect.

The fix in the change is only a partial fix. It reverts to the previous
behavior that all data needs to pass through the logging infrastructure
to
ensure its FIFO semantics, but it doesn't address the fact that the
replay
operator doesn't maintain ordering explicitly -- Timely channels aren't
guaranteed to preserve order, even if the implementation currently does.
A proper fix is to introduce a sequencer, but that needs work to not
regress in performance.

### Checklist

- [ ] This PR has adequate test coverage / QA involvement has been duly
considered. ([trigger-ci for additional test/nightly
runs](https://trigger-ci.dev.materialize.com/))
- [ ] This PR has an associated up-to-date [design
doc](https://github.com/MaterializeInc/materialize/blob/main/doc/developer/design/README.md),
is a design doc
([template](https://github.com/MaterializeInc/materialize/blob/main/doc/developer/design/00000000_template.md)),
or is sufficiently small to not require a design.
  <!-- Reference the design in the description. -->
- [ ] If this PR evolves [an existing `$T ⇔ Proto$T`
mapping](https://github.com/MaterializeInc/materialize/blob/main/doc/developer/command-and-response-binary-encoding.md)
(possibly in a backwards-incompatible way), then it is tagged with a
`T-proto` label.
- [ ] If this PR will require changes to cloud orchestration or tests,
there is a companion cloud PR to account for those changes that is
tagged with the release-blocker label
([example](https://github.com/MaterializeInc/cloud/pull/5021)).
<!-- Ask in #team-cloud on Slack if you need help preparing the cloud
PR. -->
- [ ] If this PR includes major [user-facing behavior
changes](https://github.com/MaterializeInc/materialize/blob/main/doc/developer/guide-changes.md#what-changes-require-a-release-note),
I have pinged the relevant PM to schedule a changelog post.

Signed-off-by: Moritz Hoffmann <mh@materialize.com>