Maintenance interval to rate-limit expensive tasks
Similarly to what compute has been doing for a while, introduce a
maintenance interval in the storage server to rate limit specific tasks.
At the moment, this only covers reporting frontiers to the controller. The
downside is that we might delay reporting frontier updates by the
maintenance interval.

At the moment, the maintenance interval is set to 10ms, the same as used by
compute. The largest expected frontier reporting regression is thus 10ms.

Signed-off-by: Moritz Hoffmann <mh@materialize.com>