storage/pg: produce probes during snapshotting
This commit fixes a defect in the postgres source where it would only
start producing upstream probes once it was finished with snapshotting
and started replication. The result of which being that with RLU
enabled, the entire snaphot would have to be buffered in memory because
no timestamp bindings would be produced during snapshotting, and thus no
updates could transit the reclock operator.

This fix adds a separate "statistics" operator to the postgres source,
like the mysql source already has. As in the mysql source, this operator
is responsible for issueing upstream probes, so probing is decoupled
from the replication process and can start immediately when the source
is rendered.