Scalable schema fetching from postgres
Improve the scalability of `publication_info`, which fetches schema
information from postgres. It now accepts a parameter for which schemas
it should fetch, and is rewritten to avoid fetching data in a loop.
Instead, it fetches the information for all OIDs in a single SQL query.

For pulling the schema of 1k tables, this reduces the time from about 1s
to <10ms in local tests.

Signed-off-by: Moritz Hoffmann <mh@materialize.com>