storage/pg: emit an early probe during snapshotting
This commit changes the Postgres source to emit an initial upstream
probe immediately after the creation of the replication slot. The
purpose of this is to produce the first timestamp binding (under RLU) as
early as possible, so it can be used to reclock the snapshot updates
rather than buffering them in the reclock operator.

Prior to this change, the first probe would only be emitted after the
snapshot had been fully processed, which meant that that all snapshot
updates needed to be buffered in the reclock operator, making the
Postgres source require memory proportional to the size of the snapshot.