Properly compute storage reduced count
This fixes another fix, which was correct but too conservative. The storage
protocol isn't flattened (at least until we land #31261), so we don't just
have to count the commands after compaction, but the semantic commands. For
example, `AllowCompaction` can contain main commands, and we need to count
them as such. Otherwise, we'll compact too frequently, showing in CPU
profiles when we host many storage objects.

Signed-off-by: Moritz Hoffmann <mh@materialize.com>