catalog: add wallclock lag rollup view
This commit adds a new view, `mz_internal.mz_wallclock_global_lag` that
shows the most recent wallclock lag values from all collections. The
intention is to have something that's easier to use for people that want
to inspect the current freshness of their objects and are not interested
in historical values.