Converting container builder
At the moment, we have many places where we consolidate opportunistically
to reduce the amount of data as early as possible. This works fine for
vectors because we have dedicated functions, and vectors offer mutable
access to their contents. It's much harder to implement for containers that
do not offer mutable access, such as columns. To avoid losing the benefit
of eager consolidation, we introduce a converting container builder that
wraps two container builders: The first absorbs inputs, the second absorbs
the outputs of the first. This way, we can consolidate the inputs and then
stage them in a column, combining the benefits of consolidation and
columnar data layout.

This has the beneficial effect that allocations contained in the
consolidated output do not last beyond the second container builder, i.e.,
they are short-lived and can be easily recycled through the allocator on
the same thread.

Signed-off-by: Moritz Hoffmann <mh@materialize.com>