Less restrictive flat_map_fallible
The `flat_map_fallible` function currently fixes its output container type
to vectors. This is a leftover from when we didn't support any other type,
but there is no reason to encode this restriction here. Instead, we do what
we do elsewhere: let the container builder specify the output type.

Signed-off-by: Moritz Hoffmann <mh@materialize.com>