catalog,storage: add mz_wallclock_global_lag_histogram
This commit adds a new built-in source,
`mz_wallclock_global_lag_histogram_raw` that keeps 30 days worth of data
by default. No data is filled in yet.

The `_raw` suffix suggests that the histogram counts are stored as diffs
for efficiency. This makes the histogram harder to query, so this commit
also adds a view `mz_wallclock_lag_histogram` that lifts the count into
a column.