controller: refactor wallclock lag refresh code
This commit refactors the existing controller code to refresh wallclock
lag introspection in preparation of the addition of a wallclock lag
histogram. Mostly it renames things from "wallclock lag" to "wallclock
lag history", to avoid confusion when we also add histogram updating
logic there. It also changes the structure of the
`refresh_wallclock_lag` methods a bit, to make them more similar between
the two controllers.