format_char_str: Avoid costly decoding if possible
The `format_char_str` function formats a string as a fixed-length string,
potentially cutting off overhanging white space. It needs to decode the
string to find the position of the `length` character, which is a slow
operation.

This change introduces an optimization where if a string certainly isn't
longer than the length in characters, we bail out early. The reasoning for
this is simple: a string of length $n$ can have at most $n$ characters, so
the length serves as an upper bound for the number of characters.

Related: MaterializeInc/database-issues#9125

Signed-off-by: Moritz Hoffmann <mh@materialize.com>