postgres replication: Eager string decoding
The postgres replication implementation represented rows of data as
`Vec<Option<Vec<u8>>>`, which is accurate but potentially inefficient.

Instead, we switch to `Row` containing `Datum::String` or `Datum::Null`.
This causes us to do less work in total, with a slight caveat that it is
_different_ work on the replication worker. I'm not sure if cloning Bytes
into the nested vector is more work than constructing a `Row`, but there is
a possibility for performance changes. A `Row` certainly is more compact in
memory.

Related: MaterializeInc/database-issues#9125

Fixes: MaterializeInc/database-issues#9123

Signed-off-by: Moritz Hoffmann <mh@materialize.com>