Implement `determine_timestamp` using constraints (#27815)
This PR explores what it would look like to pilot the
`determine_timestamp` functions using constraints, rather than the
current interwoven arguments with .. at least unclear-to-me semantics.
The implementation reverses out some of the intents from the settings of
`QueryWhen`, the optional oracle timestamp, the isolation level, and the
timeline, and translates them to constraints with associated "reasons".

The intent is to nudge the implementation to one that starts with actual
constraints with reasons, ideally pushing those constraints up the stack
to the moment they are expressed. There are also surprising (to me)
moments called out in the code where e.g. we avoid choosing the freshest
timestamp as a function of `QueryWhen` rather than our isolation
(presumably elsewhere the isolation results in a setting of the
`QueryWhen`?).

The logic around `StrongSessionSerializable` is interesting, and the
`Preference` enum isn't rich enough to represent it. I think this is
because it is the only logic at the moment that intentionally trades off
freshness and responsiveness; the other code paths sort of end up being
wired to do reasonable things, as a result of the contexts in which we
call the function with various arguments.

### Motivation

<!--
Which of the following best describes the motivation behind this PR?

  * This PR fixes a recognized bug.

    [Ensure issue is linked somewhere.]

  * This PR adds a known-desirable feature.

    [Ensure issue is linked somewhere.]

  * This PR fixes a previously unreported bug.

    [Describe the bug in detail, as if you were filing a bug report.]

  * This PR adds a feature that has not yet been specified.

[Write a brief specification for the feature, including justification
for its inclusion in Materialize, as if you were writing the original
     feature specification.]

   * This PR refactors existing code.

[Describe what was wrong with the existing code, if it is not obvious.]
-->

### Tips for reviewer

<!--
Leave some tips for your reviewer, like:

    * The diff is much smaller if viewed with whitespace hidden.
    * [Some function/module/file] deserves extra attention.
* [Some function/module/file] is pure code movement and only needs a
skim.

Delete this section if no tips.
-->

### Checklist

- [ ] This PR has adequate test coverage / QA involvement has been duly
considered. ([trigger-ci for additional test/nightly
runs](https://trigger-ci.dev.materialize.com/))
- [ ] This PR has an associated up-to-date [design
doc](https://github.com/MaterializeInc/materialize/blob/main/doc/developer/design/README.md),
is a design doc
([template](https://github.com/MaterializeInc/materialize/blob/main/doc/developer/design/00000000_template.md)),
or is sufficiently small to not require a design.
  <!-- Reference the design in the description. -->
- [ ] If this PR evolves [an existing `$T ⇔ Proto$T`
mapping](https://github.com/MaterializeInc/materialize/blob/main/doc/developer/command-and-response-binary-encoding.md)
(possibly in a backwards-incompatible way), then it is tagged with a
`T-proto` label.
- [ ] If this PR will require changes to cloud orchestration or tests,
there is a companion cloud PR to account for those changes that is
tagged with the release-blocker label
([example](https://github.com/MaterializeInc/cloud/pull/5021)).
<!-- Ask in #team-cloud on Slack if you need help preparing the cloud
PR. -->
- [ ] This PR includes the following [user-facing behavior
changes](https://github.com/MaterializeInc/materialize/blob/main/doc/developer/guide-changes.md#what-changes-require-a-release-note):
- <!-- Add release notes here or explicitly state that there are no
user-facing behavior changes. -->

---------

Co-authored-by: Dov Alperin <dov.alperin@materialize.com>