Turn jsonb stringify functions into separate functions
Some of the jsonb functions can be used in contexts where the output is
either jsonb or strings. Previously, we'd encode this as part of the
function call by supplying a `bool` field. This is an abstraction leak and
instead it would be better to have separate functions.

This change absorbs the `stringify` parameter into different functions. No
behavior change expected.

Signed-off-by: Moritz Hoffmann <mh@materialize.com>