doc: Adjust cluster size in Postgres source guide (#32058)
<!--
Describe the contents of the PR briefly but completely.

If you write detailed commit messages, it is acceptable to copy/paste
them
here, or write "see commit messages for details." If there is only one
commit
in the PR, GitHub will have already added its commit message above.
-->

### Motivation
The Postgres source guide initially suggests a `50cc` size but later on
talks about a `200cc` which is not mentioned in the doc.