Tag numeric as not preserving order
Imagine two numerics which cast to the same floating-point number:
with our two-part numeric encoding, the floating-point part would
compare equal, and the binary encoding might encode in whatever order.

We definitely could use numeric's ordering in some circumstances, but
we'd need more than a boolean to represent that. Just opt it out for
now.