timely-util: fix cloning `Column::Align`
The implementation of `Column::clone` for the `Align` variant would
first allocate a zero-filled region with `alloc_aligned_zeroed`, and
then attempt to _append_ the bytes from the original region to its end
using `Region::extend_from_slice`. The correct thing to do is to
`copy_from_slice` instead, overwriting the zeroes.

Furthermore, the `Column` implementation calls `copy_from_slice` on a
region returned by `alloc_aligned_zeroed` in a couple places. The
returned region can be larger than the requested length, in which case
`copy_from_slice` would panic because it expects both source and
destination to have the same length. This is fixed here by slicing the
destination to the source length.