controller: report compute introspection updates promptly
This commit fixes a bug where certain types of introspection updates
(frontiers and wallclock lag) would be reported with a 1s delay, causing
confusion particularly around comparing compute with storage frontiers.

The bug was caused by the compute controller only reporting frontier
advancements to the storage controller once per second. Frontier
introspection is also produced once per second, but by the per-instance
controllers, and the compute controller couldn't wait on the instance
controllers because it had no way to know when they finished emitting
frontier updates.

The fix implemented here is to have a "compute introspection sink" task
that continually reads from the compute controller-internal
`introspection_rx` and pushes the introspection updates to the storage
controller. This task cannot hold a reference to the storage controller,
so instead it gets initialized with the sending end of the channels
connected to the `CollectionManager`. This way it can push updates to
the persist writer tasks directly, without going through the
`StorageController`.

We need to be careful about when to start the task: During controller
initialization, the storage controller doesn't yet know about the
introspection collections, so attempting to get channels to send updates
for them won't succeed. Instead we need to wait until bootstrapping has
progressed until after the storage collections have been initialized
from the catalog, at which point we can start the introspection sink
task.