Self hosted bump terraform module versions (#32096)
<!--
Describe the contents of the PR briefly but completely.

If you write detailed commit messages, it is acceptable to copy/paste
them
here, or write "see commit messages for details." If there is only one
commit
in the PR, GitHub will have already added its commit message above.
-->

### Motivation

Bumping the Terraform modules versions for all cloud providers to the
latest ones. With the latest versions, we have:
- TLS enabled by default for all
- NLB support enabled by default for all
- lgalloc support for AWS enabled by default

---------

Co-authored-by: Dennis Felsing <dennis@felsing.org>