Disable backtrace collection in `anyhow`
The `anyhow` crate automatically collects backtraces when creating
`anyhow::Error`s if the `RUST_BACKTRACE=1` env variable is set, and uses
them to enhance the errors' Debug representation. This is a feature we
don't actively rely on but that comes with significant costs:

 * Collecting backtraces significantly increases the memory and CPU
   costs of creating error values, which can be both expensive and
   surprising.
 * In rare cases, collecting backtraces can deadlock the process, if
   jemalloc decides to collect a profiling sample at the same time.

This commit therefore disables backtrace collection in `anyhow` by
setting `RUST_LIB_BACKTRACE=0`.

This also disables backtrace collection in all other libraries that use
`std::backtrace::Backtrace::capture`, but I'm not aware of any we use.
Sentry does collect backtraces, but it uses the `backtrace` crate, which
isn't affected by `RUST_LIB_BACKTRACE`.