Compute: Maintain row arena while fulfilling peek
Maintain a single row arena allocation while collecting the data for a
peek. We clear the arena after every value to ensure bounded size.

This does not preserve the contents of the arena, so most of the comment
in the code still applies. Preserving contents would be an option, but
would require an interface change to the arena.

Signed-off-by: Moritz Hoffmann <mh@materialize.com>