controller: emit undefined lag for empty collections
This commit changes the wallclock lag measurement logic in the two
controllers to emit "undefined" lag measurements for collections that
are empty, i.e., contain no readable times. The intention is to avoid
glitches where wallclock lag for unhydrated collections is nonsensical
(e.g. 55 years).

The "undefined" value is represented differently for introspection and
for Prometheus metrics. For introspection, we can use the SQL `NULL`
value, but Prometheus metrics values are integers, so we have to use
`u64::MAX` instead.