Persist source: retain row allocation
At the moment, the persist source would only retain a row allocation if it
had an MFP. The MFP needs to write into a different allocation, so it makes
sense to recycle the allocation here. The non-MFP path didn't retain the
allocation. This is suboptimal because subsequently we have to right-size
another row allocation to decode into, which can involve $log n$
reallocations, where $n$ is the size of the row. Instead, clone the output
and retain the probably right-sized row allocation.

Signed-off-by: Moritz Hoffmann <mh@materialize.com>