*: don't report caught panics to Sentry
Sentry's default panic handler runs before the "enhanced" panic handler
we install. So if in the enhanced handler we decide that the panic
should be caught instead of crashing the process, we have already
reported it to Sentry. This has been causing incident reports and
confusion in the past.

This commit fixes the issue by disabling Sentry's default panic
integration and instead running the Sentry panic handler manually as
part of the enhanced panic handler _after_ it has checked whether the
panic has been caught.