compute,storage: reduce the size of cluster commands
`ComputeCommand` and `StorageCommand` are enums, so their size is
dependent on the size of the largest variant. Both contain some huge
variants, in particular `UpdateConfiguration` and the ones describing
dataflows to install, bringing the size of the enums to 3-4 KB. In
contrast, the variant we send most frequently, `AllowCompaction`, is
only 40 bytes in size, so when handling an `AllowCompaction` command, we
waste 3-4 KB of space. Commands are stored in histories and channels, so
this amount of waste can have an impact on our memory usage.

This commit removes the waste and brings the size of the command enum
down to 40 bytes by boxing large fields. This makes code creating
commands a bit more noisy, but no way to avoid that I think.