controller: consolidate wallclock lag refresh intervals
This commit merges the two feature flags controlling the intervals in
which updates for wallclock lag history and histogram introspection
should be written to persist. Having two provides little value, so we
can get rid of some clutter.

This commit also renames the "refresh interval" to "recording interval",
to make it clearer that it's about writing updates to persist, as
opposed to refreshing the in-memory tracking state (which is done every
second).