FIX: use the optimized version when incrementing a metric
This was introduced in a recent refactoring... when
we can prove that a part is equivalent with a simpler zero-column
version we swap that in, but were still generating stats based on the
original data. This fixes that, and tidies up the logic to make it
hopefully slightly more clear.