compute: identify LIR mappings by export IDs
Previously, during dataflow rendering we would emit LIR mappings
together with the ID of the `objects_to_build` object that contained the
LIR node. This is often the ID of the view over which an index is
created, or some dummy ID in the case of MVs, rather than the ID of the
object for which we install the dataflow. It caused potential confusion,
as well as the need for a new relation, `mz_dataflow_global_ids`, that
maps `objects_to_build` IDs to the dataflow that contain them.

This commit changes things to instead identify LIR mappings by the
dataflow export IDs, i.e. the IDs of the objects (indexes, MVs) for
which we install the dataflow. This will allow us to simplify the code a
bit and remove `mz_dataflow_global_ids` in favor of
`mz_compute_exports`.