ore: report panics to Sentry before aborting
The enhanced panic handler had a bug where it would report the panic to
Sentry only after spawning a thread that could abort the process before
the Sentry call had a chance to run, causing panics to be missing from
Sentry. The fix is to do the Sentry reporting earlier.