repr: store thread-local shared row inline
This PR removes the Rc indirection for the shared row and simplifies the
API by not requiring the caller to explicitly mutably borrow the
RefCell.

The constructor of the `SharedRow` is documented to panic if more than
one instance is constructed. This allows us to use a single
`Cell<Option<Row>>` as the contents of the thread local that gets
replaced with `None` when someone gets the shared row and is put back
when the `SharedRow` value is dropped.

Signed-off-by: Petros Angelatos <petrosagg@gmail.com>