dbt-materialize: release `v1.9.5` (#32383)
<!--
Describe the contents of the PR briefly but completely.

If you write detailed commit messages, it is acceptable to copy/paste
them
here, or write "see commit messages for details." If there is only one
commit
in the PR, GitHub will have already added its commit message above.
-->

### Motivation

Preparing this release to ship the new support for `retain_history`
config for materialized views
https://github.com/MaterializeInc/materialize/pull/32376