storage: document and uphold upsert correctness property
This PR fixes the correctness issue observed during concurrent execution
of two upsert sources on multi-replica clusters.

The issue stems from a violation of a correctness proprety of the
`drain_staged_input` function, that this PR now documents.

The diff is meant to be minimal to aid quick review and backporting but
it does not make any progress towards simplifying the overal logic,
which will be subsequent work.

## Explanation

When `drain_staged_input` is called in `DrainStyle::ToUpper` mode it
drains all the upsert commands that are exactly at the timestamp of
`persist_upper` and for each key it:

1. Keeps the command with the highest offset
2. Retrieves the current state from RocksDB
3. Produces a retraction of the current value if it was set
4. Produces an insertion of the new value.

In this mode the RocksDB state is NOT modified. Therefor if the function
is called twice with the same `persist_upper` and processes a command
for the same key it will produce a double retraction, since the second
call will not be aware of the first call.

The code was previously set up to call `drain_staged_input` once per
held capability with the same `persist_upper`. Due to how the rest of
the code was setup the set of held capabilities always turned out to be
a singleton when a single replica was involved, therefore not triggering
the bug.

With multiple replicas however one instance of the operator could
receive the following in its command stream:

* cap=1, batch = `[ ({key=b, value=NULL}, ts=1), ({key=a, value=a}, ts=2) ]`
* cap=2, batch = `[ ({key=a, value=b}, ts=2) ]`
* progress event to {3}

Then another instance of the operator seeing the same stream of commands
processes everything till `{2}` and manages to commit to persist.

Finally the original operator receives a persist upper progress to `{2}`
and ends up calling `drain_staged_input` twice since it holds onto two
capabilities and ends up processing the same key (`key=a`) twice.

Fixes MaterializeInc/database-issues#9160

Signed-off-by: Petros Angelatos <petrosagg@gmail.com>