storage/pg: only report snapshot size error in strict mode
Our default snapshot size determination strategy is to use the table
statistics in pg instead of performing a `SELECT COUNT(*)` on the table.

This saves resources from the upstream database since it only reads a
couple of rows from the statistics table at the expense of accuracy.

This PR silences an error that only makes sense when the `COUNT(*)`
strategy is used.

Signed-off-by: Petros Angelatos <petrosagg@gmail.com>