compute: factor out PeekResultIterator, refactor peek fulfillment
The original motivation for this is so that the code that extracts peek
results can be re-used in
https://github.com/MaterializeInc/database-issues/issues/9180, where we
want to use a different transport for sending back peek responses but
still need to read them out of arrangements the same way.

The nice side effect is that we separate extracting the result from the
logic that accumulates it in a response for sending it back. Which leads
to clearer separation.