compute: simplify passing of initial replica config
This commit simplifies how the initial replica config parameters
(arrangement_exert_effort + the zero-copy config) are passed to the
replica task inside the controller. Instead of having a separate
`InitialComputeParameters` type and setters, we can simply read the
values from the dyncfg when creating the `ReplicaConfig`.

Note that the replica task also has access to the dyncfg but we can't
read the values dynamically in `specialize_command` because they are
written to the `TimelyConfig` which needs to be stable across replica
reconnects.