ore: fix reporting of panics to Sentry
It turns out that calling `sentry_panic::panic_handler` doesn't work
because it expects the `PanicIntegration` to be enabled and silently
does nothing when that isn't the case. We don't want to enable the panic
integration because that would also install Sentry's panic hook. So
instead we have to implement the panic handler logic manually.