compute: avoid frontier regressions for introspection collections
Compute introspection collections are weird in a couple ways. One of
them is that they are implicitly created by the `CreateInstance`
command, which doesn't contain a prescribed as-of. The replica always
initializes them to the minimum frontier. So if the controller
initializes its tracking for these collections to some other as-of, like
the global read frontier of the index, it might observe unexpected
frontier regressions and log noisy errors about them. The fix is to
initializes the controller's tracking state to the minimum frontier as
well.