ColumnName wraps boxed string (#32591)
ColumnName currently owns a string, but doesn't need mutable access in
all but one places. This change alters it to a boxed str, which reduces
the size from 24 to 16 bytes. It also adds a deref implementation, which
allows us to remove some of the `as_str` calls.

No urgency to merge it, just capturing what I found in my local changes.

Signed-off-by: Moritz Hoffmann <mh@materialize.com>